# knowledge/pkg/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Knowledge API server for GPTStudio.",
        "title": "Knowledge API",
        "contact": {
            "name": "Acorn Labs Inc."
        },
        "version": "1"
    },
    "paths": {
        "/datasets": {
            "get": {
                "description": "List all datasets",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "datasets"
                ],
                "summary": "List all datasets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.Dataset"
                            }
                        }
                    }
                }
            }
        },
        "/datasets/create": {
            "post": {
                "description": "Create a new dataset",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "datasets"
                ],
                "summary": "Create a new dataset",
                "parameters": [
                    {
                        "description": "Dataset object",
                        "name": "dataset",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Dataset"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.Dataset"
                        }
                    }
                }
            }
        },
        "/datasets/{id}": {
            "get": {
                "description": "Get a dataset by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "datasets"
                ],
                "summary": "Get a dataset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dataset ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.Dataset"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a dataset by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "datasets"
                ],
                "summary": "Delete a dataset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dataset ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/datasets/{id}/documents/{doc_id}": {
            "delete": {
                "description": "Remove a document from a dataset by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "datasets"
                ],
                "summary": "Remove a document from a dataset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dataset ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Document ID",
                        "name": "doc_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/datasets/{id}/files/{file_id}": {
            "delete": {
                "description": "Remove a file from a dataset by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "datasets"
                ],
                "summary": "Remove a file from a dataset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dataset ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "File ID",
                        "name": "file_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/datasets/{id}/ingest": {
            "post": {
                "description": "Ingest content into a dataset by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "datasets"
                ],
                "summary": "Ingest content into a dataset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dataset ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.IngestResponse"
                        }
                    }
                }
            }
        },
        "/datasets/{id}/retrieve": {
            "post": {
                "description": "Retrieve content from a dataset by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "datasets"
                ],
                "summary": "Retrieve content from a dataset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dataset ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/vectorstore.Document"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "gin.H": {
            "type": "object",
            "additionalProperties": {}
        },
        "types.Dataset": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "description": "Dataset ID - must be a valid RFC 1123 hostname",
                    "type": "string",
                    "format": "hostname_rfc1123",
                    "example": "asst-12345"
                }
            }
        },
        "types.IngestResponse": {
            "type": "object",
            "properties": {
                "documents": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "vectorstore.Document": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "similarity_score": {
                    "type": "number"
                }
            }
        }
    }
}